cleaning up commands
